requirements.txt
flask
pandas
thefuzz
python-Levenshtein
groq
python-dotenv
requests
gunicorn
